last layer len from config
